refactor(ast)!: take an i32 in Expr::int and Lit::int
`isize` depends on the platform, while the values these helpers are for,
hex and octal literals, always fit in 32 bits. Decimal literals are kept as
text in `Lit::Int` and are not affected.

BREAKING CHANGE: `Expr::int` and `Lit::int` take an `i32` instead of an
`isize`.

+++ src/parser/ast.rs
@@ -229,7 +229,7 @@ impl Expr {
         ExprKind::Ident(name.into()).into()
     }
 
-    pub fn int(i: isize) -> Self {
+    pub fn int(i: i32) -> Self {
         ExprKind::Literal(Lit::Int(i.to_string())).into()
     }
 
@@ -573,7 +573,7 @@ impl Lit {
         Lit::Str(s.into())
     }
 
-    pub fn int(i: isize) -> Self {
+    pub fn int(i: i32) -> Self {
         Lit::Int(i.to_string())
     }
 }
